requirements.txt
-i https://pypi.org/simple
adal==1.2.1
aiohttp==3.6.2
aiounittest==1.3.0
async-timeout==3.0.1
attrs==20.3.0
babel==2.7.0
botbuilder-core==4.11.0
botbuilder-dialogs==4.11.0
botbuilder-integration-aiohttp==4.11.0
botbuilder-schema==4.11.0
botframework-connector==4.11.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==2.8
datedelta==1.3
emoji==0.6.0
grapheme==0.6.0
http-noah[async]==0.1.10
idna==2.10
isodate==0.6.0
jsonpickle==1.2
msal==1.2.0
msrest==0.6.10
multidict==4.7.6
multipledispatch==0.6.0
oauthlib==3.1.0
pycparser==2.20
pydantic==1.7.2
pyjwt[crypto]==1.5.3
python-dateutil==2.8.1
pytz==2020.4
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text==1.0.2a2
regex==2019.8.19
requests-oauthlib==1.3.0
requests==2.23.0
six==1.15.0
structlog==20.1.0
urllib3==1.25.11
yarl==1.6.3
